Remove SqlDiffView
